Refactored celery tasks to be stored in the recipe not a global variable
