+++ backend/recipes/base.py
@@ -3,7 +3,7 @@ Common code for recipes, handles running recipe files.
 
 The recipe steps are defined in a json file and they can be strung together with "goto" statements.
 Each recipe step can also call a function that changes the desired state of the hardware. These methods are defined
-at the bottom of this file and the parameters they accept can be seen in the "baseTask" definition in the
+in the tasks.py file and the parameters they accept can be seen in the "baseTask" definition in the
 documentation of the plan object below.
 
 You'll probably want to take a look at sample recipe in the recipes.files package. This will
@@ -48,7 +48,7 @@ plan object
                                 The nr id of the step to execute if the
                                 user picks this option.
                 baseTask
-                    Name of the function to execute in the recipes.base module.
+                    Name of the function to execute in the recipes.tasks module.
                     One of:
                         heat
                             parameters: {'temp': 65}
@@ -87,6 +87,7 @@ class Recipe:
     icon = ''
     time = ''
     currentRecipe = None
+    currentTask = None
     mutex = threading.Lock()
 
     def __init__(self, plan):
@@ -122,7 +123,7 @@ class Recipe:
         self.time = ''
         hardware.turnHeaterOff()
         hardware.turnCoolerOff()
-        celery.stopTask()
+        self.stopTask()
 
     def getStatus(self):
         """
@@ -169,7 +170,7 @@ class Recipe:
         """
         self.mutex.acquire()
         if self.status == 'running':
-            if celery.isTaskComplete():
+            if self.isTaskComplete():
                 currentStep = self.plan['steps'][self.step]
                 if('done' in currentStep) and (currentStep['done'] == True):
                     self.stop()
@@ -238,20 +239,9 @@ class Recipe:
             self.icon = step['icon']
 
         if 'baseTask' in step and step['baseTask'] != 'humanTask':
-            if ('task' in step):
-                task = step['task']
-                base = False
-            else:
-                task = step['baseTask']
-                base = True
-
-            if celery.runTask(task, step['parameters'], base):
-                self.status = 'running'
-            else:
-                self.status = 'error'
-                message = 'Internal error. Task already running.'
-                return False
-
+            self.currentTask = celery.runTask(step['baseTask'], step['parameters'])
+            self.status = 'running'
+            
         if('parameters' in step) and ('time' in step['parameters']):
             self.time = step['parameters']['time']
         else:
@@ -263,153 +253,27 @@ class Recipe:
 
         return True
 
+    def isTaskComplete(self):
+        """
+        Check if the currently running task has completed.
+        :return:
+        True
+            Task is complete.
+        False
+            Task is still running.
+        """
+        if self.currentTask is not None:
+            return self.currentTask.ready()
 
-def heat(parameters):
-    """
-    Turn on the heater and reach a target temperature.
+        return True
 
-    :param parameters:
-        dictionary
-            'temp' - the desired temperature
-    :return:
-        None
-    """
-    targetTemp = parameters['temp']
-    celery.logger.info('heating water to ' + str(targetTemp) + '...')
-    hardware.turnHeaterOn()
-    while hardware.getTemp() < targetTemp:
-        hardware.sleep(0.5)
-    hardware.turnHeaterOff()
+    def stopTask(self):
+        """
+        Stop the currently running task.
+        :return:
+            None
+        """
+        if self.currentTask is not None:
+            self.currentTask.revoke(terminate=True)
+            self.currentTask = None
 
-
-def cool(parameters):
-    """
-    Turn on the cooler and reach a target temperature.
-
-    :param parameters:
-        dictionary
-            'temp' - the desired temperature
-    :return:
-        None
-    """
-    targetTemp = parameters['temp']
-    celery.logger.info('cooling water to ' + str(targetTemp) + '...')
-    hardware.turnCoolerOn()
-    while hardware.getTemp() > targetTemp:
-        hardware.sleep(0.5)
-    hardware.turnCoolerOff()
-
-
-def maintainCool(parameters):
-    """
-    Maintain a certain temperature using the cooler for a specified amount of time.
-
-    :param parameters:
-        dictionary
-            'temp' - the desired temperature
-            'tolerance' - how much above or below the desired temperature to let
-                            the temperature drift before turning on cooler again
-            'time' - the amount of time to maintain the temperature in seconds
-    :return:
-        None
-    """
-    parameters['type'] = 'cool'
-    maintain(parameters)
-
-
-def maintainHeat(parameters):
-    """
-    Maintain a certain temperature using the heater for a specified amount of time.
-
-    :param parameters:
-        dictionary
-            'temp' - the desired temperature
-            'tolerance' - how much above or below the desired temperature to let
-                            the temperature drift before turning on heater again
-            'time' - the amount of time to maintain the temperature in seconds
-    :return:
-        None
-    """
-    parameters['type'] = 'heat'
-    maintain(parameters)
-
-
-def maintain(parameters):
-    """
-    Maintain a certain temperature using either the cooler or heater for a specified amount of time.
-
-    :param parameters:
-        dictionary
-            'temp' - the desired temperature
-            'tolerance' - how much above or below the desired temperature to let
-                            the temperature drift before turning on the heater/cooler again
-            'time' - the amount of time to maintain the temperature in seconds
-            'type' - one of:
-                - heat
-                    Maintain temperature using heater.
-                - cool
-                    Maintain temperature using cooler.
-    :return:
-        None
-    """
-    duration = parameters['time']
-    targetTemp = parameters['temp']
-    tolerance = parameters['tolerance']
-    type = parameters['type']
-
-    interval = 0.5
-    start = hardware.secondSinceStart()
-
-    while (hardware.secondSinceStart() - start) < duration:
-        hardware.sleep(interval)
-        currentTemp = hardware.getTemp()
-        celery.logger.info('temperature @ ' + str(currentTemp))
-        if currentTemp - tolerance > targetTemp:
-            if type == 'heat':
-                hardware.turnHeaterOff()
-            else:
-                hardware.turnCoolerOn()
-        if currentTemp + tolerance < targetTemp:
-            if type == 'heat':
-                hardware.turnHeaterOn()
-            else:
-                hardware.turnCoolerOff()
-
-    hardware.turnHeaterOff()
-    hardware.turnCoolerOff()
-
-
-def pump(parameters):
-    """
-    Dispense a certain amount of liquid from a pump.
-
-    :param parameters:
-        dictionary
-            'pump' - one of: 'A' or 'B'
-            'volume' - volume to dispense in ml
-    :return:
-        None
-    """
-    pump = parameters['pump']
-    volume = parameters['volume']
-    hardware.pumpDispense(pump, volume)
-
-
-def stir(parameters):
-    """
-    Turn on the stirrer for a predefined amount of time.
-
-    :param parameters:
-        dictionary
-            'time' - the amount of time to turn on the stirrer for
-    :return:
-        None
-    """
-    duration = parameters['time']
-
-    interval = 0.5
-    start = hardware.secondSinceStart()
-    hardware.turnStirrerOn()
-    while (hardware.secondSinceStart() - start) < duration:
-        hardware.sleep(interval)
-    hardware.turnStirrerOff()
